pyproject.toml
[project]
name = "assignment-1-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=5.44.1",
    "openai-agents>=0.2.10",
    "tavily-python>=0.7.11",
]
